Moving home component out of common directory
