+++ frontend/src/components/home/midsection/lower_mid.jsx
