+++ frontend/src/components/home/midsection/mid.css
