+++ frontend/src/components/home/midsection/upper_mid.jsx
